feat: Moved UserProfileImagePicker from ads folder (#17111)
* fix: Moved UserProfileImagePicker from ads folder

* design system package version update
